// web3-game/typechain-types/ClickerGame.ts
export interface ClickerGameInterface extends Interface {
  getFunction(
    nameOrSignature: "getPlayer" | "getScore" | "players" | "updateScore"
  ): FunctionFragment;

  getEvent(nameOrSignatureOrTopic: "ScoreUpdated"): EventFragment;

  encodeFunctionData(
    functionFragment: "getPlayer",
    values: [AddressLike]
  ): string;
  encodeFunctionData(
    functionFragment: "getScore",
    values: [AddressLike]
  ): string;
  encodeFunctionData(
    functionFragment: "players",
    values: [AddressLike]
  ): string;
  encodeFunctionData(
    functionFragment: "updateScore",
    values: [BigNumberish]
  ): string;

  decodeFunctionResult(functionFragment: "getPlayer", data: BytesLike): Result;
  decodeFunctionResult(functionFragment: "getScore", data: BytesLike): Result;
  decodeFunctionResult(functionFragment: "players", data: BytesLike): Result;
  decodeFunctionResult(
    functionFragment: "updateScore",
    data: BytesLike
  ): Result;
}

export namespace ScoreUpdatedEvent {
  export type InputTuple = [player: AddressLike, newScore: BigNumberish];
  export type OutputTuple = [player: string, newScore: bigint];
  export interface OutputObject {
    player: string;
    newScore: bigint;
  }
  export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>;
  export type Filter = TypedDeferredTopicFilter<Event>;
  export type Log = TypedEventLog<Event>;
  export type LogDescription = TypedLogDescription<Event>;
}

export interface ClickerGame extends BaseContract {
  connect(runner?: ContractRunner | null): ClickerGame;
  waitForDeployment(): Promise<this>;

  interface: ClickerGameInterface;

  queryFilter<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;
  queryFilter<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;

  on<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  on<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  once<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  once<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  listeners<TCEvent extends TypedContractEvent>(
    event: TCEvent
  ): Promise<Array<TypedListener<TCEvent>>>;
  listeners(eventName?: string): Promise<Array<Listener>>;
  removeAllListeners<TCEvent extends TypedContractEvent>(
    event?: TCEvent
  ): Promise<this>;

  getPlayer: TypedContractMethod<
    [_player: AddressLike],
    [[bigint, bigint] & { score: bigint; lastUpdate: bigint }],
    "view"
  >;

  getScore: TypedContractMethod<[_player: AddressLike], [bigint], "view">;

  players: TypedContractMethod<
    [arg0: AddressLike],
    [[bigint, bigint] & { score: bigint; lastUpdate: bigint }],
    "view"
  >;

  updateScore: TypedContractMethod<
    [_score: BigNumberish],
    [void],
    "nonpayable"
  >;

  getFunction<T extends ContractMethod = ContractMethod>(
    key: string | FunctionFragment
  ): T;

  getFunction(
    nameOrSignature: "getPlayer"
  ): TypedContractMethod<
    [_player: AddressLike],
    [[bigint, bigint] & { score: bigint; lastUpdate: bigint }],
    "view"
  >;
  getFunction(
    nameOrSignature: "getScore"
  ): TypedContractMethod<[_player: AddressLike], [bigint], "view">;
  getFunction(
    nameOrSignature: "players"
  ): TypedContractMethod<
    [arg0: AddressLike],
    [[bigint, bigint] & { score: bigint; lastUpdate: bigint }],
    "view"
  >;
  getFunction(
    nameOrSignature: "updateScore"
  ): TypedContractMethod<[_score: BigNumberish], [void], "nonpayable">;

  getEvent(
    key: "ScoreUpdated"
  ): TypedContractEvent<
    ScoreUpdatedEvent.InputTuple,
    ScoreUpdatedEvent.OutputTuple,
    ScoreUpdatedEvent.OutputObject
  >;

  filters: {
    "ScoreUpdated(address,uint256)": TypedContractEvent<
      ScoreUpdatedEvent.InputTuple,
      ScoreUpdatedEvent.OutputTuple,
      ScoreUpdatedEvent.OutputObject
    >;
    ScoreUpdated: TypedContractEvent<
      ScoreUpdatedEvent.InputTuple,
      ScoreUpdatedEvent.OutputTuple,
      ScoreUpdatedEvent.OutputObject
    >;
  };
}